requirements.txt
numpy==1.16.2
pandas==0.24.1
SQLAlchemy==1.3.0
mysql-connector==2.1.4
